+++ src/features/commands/index.ts
@@ -11,6 +11,7 @@ import { FcImporter } from './fcImporter';
 import { CfgYamlCommands } from './cfgyaml';
 import { PatchTester } from './patchTester';
 import { RunTests } from './runTests';
+import { CheckLoca } from './checkLoca';
 
 export function registerCommands(context: vscode.ExtensionContext) {
 	context.subscriptions.push(...FcConverter.register(context));
@@ -25,4 +26,5 @@ export function registerCommands(context: vscode.ExtensionContext) {
 	context.subscriptions.push(...CfgYamlCommands.register(context));
 	context.subscriptions.push(...PatchTester.register(context));
   context.subscriptions.push(...RunTests.register(context));
+	context.subscriptions.push(...CheckLoca.register(context));
 }