+++ src/test/suite/checkLoca.test.ts
@@ -1,0 +1,27 @@
+import * as assert from 'assert';
+import * as path from 'path';
+import * as fs from 'fs';
+import * as utils from '../../other/utils';
+import * as xml2js from 'xml2js';
+import * as vscode from 'vscode';
+
+suite('file conversion tests', () => {
+  test('check localization commands', async () => {
+    const relTestPath = 'test/suite/data/loca';
+    const tempTestPath = path.resolve('../../out/' + relTestPath);
+
+    utils.ensureDir(tempTestPath);
+    fs.copyFileSync(path.join('../../src/', relTestPath, 'texts_english.xml'), path.join(tempTestPath, 'texts_english.xml'));
+    fs.copyFileSync(path.join('../../src/', relTestPath, 'texts_chinese.xml'), path.join(tempTestPath, 'texts_chinese.xml'));
+
+    const modifierPath = path.join(tempTestPath, 'texts_english.xml');
+    const resultPath = path.join(tempTestPath, 'texts_chinese.xml');
+    await vscode.commands.executeCommand('anno-modding-tools.addMissingLoca', vscode.Uri.file(modifierPath));
+    assert(fs.existsSync(resultPath));
+
+    const resultContent = /*await xml2js.parseStringPromise*/(fs.readFileSync(resultPath, 'utf8'));
+    const expectedContent = /*await xml2js.parseStringPromise*/(fs.readFileSync(path.join('../../src/', relTestPath, 'texts_chinese-expectation.xml'), 'utf8'));
+
+    assert.deepStrictEqual(resultContent, expectedContent);
+  });
+});